refactor: move CLI input helpers to new module
